requirements.txt
aiohttp
aiogram>=3.0.0
emoji
pandas
google-generativeai